backend/requirements.txt
anyio==3.7.1
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
click==7.1.2
colorama==0.4.6
cryptography==41.0.2
dnspython==2.4.0
Flask==1.1.4
Flask-Cors==4.0.0
Flask-JWT-Extended==4.2.3
h11==0.14.0
httpcore==0.17.3
idna==3.4
itsdangerous==1.1.0
Jinja2==2.11.3
jwt==1.3.1
MarkupSafe==2.1.3
pycparser==2.21
PyJWT==2.8.0
pymongo==4.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
sniffio==1.3.0
watchdog==3.0.0
Werkzeug==1.0.1
